+++ core/vil1/file_formats/vil1_iris.cxx
@@ -247,7 +247,7 @@ bool vil1_iris_generic_image::write_header()
 
 vil1_image vil1_iris_generic_image::get_plane(unsigned int plane) const
 {
-  assert(plane < planes_);
+  assert((int)plane < planes_);
   vcl_cerr << __FILE__ ": do something for vil1_iris_generic_image::get_plane\n";
   return 0;
 }
